take 2 at guice fixes
